+++ src/main/java/com/ligg/service/VideoService.java
@@ -8,5 +8,5 @@ public interface VideoService {
     // 查询所有视频
     List<Video> list();
     // 根据id查询视频
-    Video findById(Integer id);
+    Video findById(Long id);
 }
